Create a sparse confrontation query

Create a sparse confrontation query. Only errors and missing are stored. This can be useful alternative to confront_tbl() which stores all results of a tbl validation in a table with length(rules) columns and nrow(tbl) rows. Note that the result of this function is a (lazy) query object that still needs to be executed in the database, e.g. with dplyr::collect(), dplyr::collapse() or dplyr::compute().

confront_tbl_sparse(tbl, x, key = NULL, union_all = TRUE, check_rules = TRUE)

Arguments

tbl
tbl_sql object, table in a database, retrieved with dplyr::tbl()

x
validate::validator() object with validation rules.

key
character with key column name.

union_all
if FALSE each rule is a separate query.

check_rules
if TRUE it is checked which rules 'work' on the db.

Value

A object with the necessary information: see details

Details

The return value of the function is a list with:

$query: A dbplyr::tbl_dbi() object that refers to the confrontation query.

$errors: The validation rules that are not working on the database

$working: A logical with which expression are working on the database.

$exprs: All validation expressions.

See also

Examples

# create a table in a database
income <- data.frame(id = letters[1:2], age=c(12,35), salary = c(1000,NA))
con <- dbplyr::src_memdb()
tbl_income <- dplyr::copy_to(con, income, overwrite=TRUE)
print(tbl_income)

#> # Source:   table<income> [?? x 3]
#> # Database: sqlite 3.34.1 [:memory:]
#>   id      age salary
#>   <chr> <dbl>  <dbl>
#> 1 a        12   1000
#> 2 b        35     NA

# Let's define a rule set and confront the table with it:
rules <- validator( is_adult   = age >= 18
                  , has_income = salary > 0
                  , mean_age   = mean(age,na.rm=TRUE) > 20
                  )

# and confront!
cf <- confront(tbl_income, rules)
print(cf)

#> Object of class 'tbl_validation'
#> Call:
#>     confront.tbl_sql(tbl = dat, x = x, ref = ref, key = key, sparse = sparse)
#> 
#> Confrontations: 3
#> Tbl           : income (:memory:)
#> Sparse        : FALSE
#> Fails         : [??] (see `values`, `summary`)
#> Errors        : 0

summary(cf)

#>         name items npass nfail nNA warning error                   expression
#> 1   is_adult     2     1     1   0   FALSE FALSE         (age - 18) >= -1e-08
#> 2 has_income     2     1     0   1   FALSE FALSE                   salary > 0
#> 3   mean_age     1     1     0   0   FALSE FALSE mean(age, na.rm = TRUE) > 20

# Values (i.e. validations on the table) can be retrieved like in `validate` 
# with`type="matrix"` (simplify = TRUE)
values(cf, type = "matrix")

#> [[1]]
#>      is_adult has_income
#> [1,]    FALSE       TRUE
#> [2,]     TRUE         NA
#> 
#> [[2]]
#>      mean_age
#> [1,]     TRUE
#> 

# But often this seems more handy:
values(cf, type = "tbl")

#> # Source:   lazy query [?? x 3]
#> # Database: sqlite 3.34.1 [:memory:]
#>   is_adult has_income mean_age
#>      <int>      <int>    <int>
#> 1        0          1        1
#> 2        1         NA        1

# We can see the sql code by using `show_query`:
show_query(cf)

#> <SQL>
#> SELECT (`age` - 18.0) >= -1e-08 AS `is_adult`, `salary` > 0.0 AS `has_income`, AVG(`age`) OVER () > 20.0 AS `mean_age`
#> FROM `income`

# identical
show_query(values(cf, type = "tbl"))

#> <SQL>
#> SELECT (`age` - 18.0) >= -1e-08 AS `is_adult`, `salary` > 0.0 AS `has_income`, AVG(`age`) OVER () > 20.0 AS `mean_age`
#> FROM `income`

# adding a key often is handy in a database
cf <- confront(tbl_income, rules, key = "id")
print(cf)

#> Object of class 'tbl_validation'
#> Call:
#>     confront.tbl_sql(tbl = dat, x = x, ref = ref, key = key, sparse = sparse)
#> 
#> Confrontations: 3
#> Tbl           : income (:memory:)
#> Key column    : id
#> Sparse        : FALSE
#> Fails         : [??] (see `values`, `summary`)
#> Errors        : 0

values(cf, type="tbl")

#> # Source:   lazy query [?? x 4]
#> # Database: sqlite 3.34.1 [:memory:]
#>   id    is_adult has_income mean_age
#>   <chr>    <int>      <int>    <int>
#> 1 a            0          1        1
#> 2 b            1         NA        1

# sparse results in db
cf_sparse <- confront(tbl_income, rules, sparse=TRUE)
values(cf_sparse, type="tbl")

#> # Source:   lazy query [?? x 3]
#> # Database: sqlite 3.34.1 [:memory:]
#>     row rule        fail
#>   <int> <chr>      <int>
#> 1     1 is_adult       1
#> 2     2 has_income    NA