Aggregate a validation

Create either statistics by rule or record.

# S3 method for tbl_validation
aggregate(x, by = c("rule", "record"), ...)

Arguments

x
tbl_validation object

by
either by "rule" or by "record"

...
not used

Details

The result of a confront() on a db tbl results in a lazy query. That is it builds a query without executing it. To store the result in the database use compute() or values().

Examples

income <- data.frame(id = 1:2, age=c(12,35), salary = c(1000,NA))
con <- dbplyr::src_memdb()
tbl_income <- dplyr::copy_to(con, income, overwrite=TRUE)
print(tbl_income)

#> # Source:   table<income> [?? x 3]
#> # Database: sqlite 3.33.0 [:memory:]
#>      id   age salary
#>   <int> <dbl>  <dbl>
#> 1     1    12   1000
#> 2     2    35     NA

# Let's define a rule set and confront the table with it:
rules <- validator( is_adult   = age >= 18
                    , has_income = salary > 0
)

# and confront!
# in general with a db table it is handy to use a key
cf <- confront(tbl_income, rules, key="id")
aggregate(cf, by = "rule")

#> # Source:   lazy query [?? x 4]
#> # Database: sqlite 3.33.0 [:memory:]
#>   rule       npass nfail   nNA
#>   <chr>      <int> <int> <int>
#> 1 is_adult       1     1     0
#> 2 has_income     1     0     1

aggregate(cf, by = "record")

#> # Source:   lazy query [?? x 3]
#> # Database: sqlite 3.33.0 [:memory:]
#>      id nfails   nNA
#>   <int>  <dbl> <dbl>
#> 1     1      1     0
#> 2     2      0     1

# to tweak performance of the db query the following options are available
# 1) store validation result in db
cf <- confront(tbl_income, rules, key="id", compute = TRUE)
# or identical
cf <- confront(tbl_income, rules, key="id")
cf <- compute(cf)

# 2) Store the validation sparsely
cf <- confront(tbl_income, rules, key="id", sparse=TRUE )

show_query(compute(cf))

#> <SQL>
#> SELECT *
#> FROM `dbplyr_004`

values(cf, type="tbl")

#> # Source:   table<dbplyr_004> [?? x 3]
#> # Database: sqlite 3.33.0 [:memory:]
#>      id rule        fail
#>   <int> <chr>      <int>
#> 1     1 is_adult       1
#> 2     2 has_income    NA