Confront tbl_sql with validator rules.

Confront tbl_sql objects with validator rules. This function makes it possible to execute validator() rules on database tables.

confront.tbl_sql(tbl, x, ref, key = NULL, sparse = FALSE, compute = FALSE, ...)

Arguments

tbl
tbl_sql object, table in a database, retrieved with dplyr::tbl()

x
validate::validator() object with validation rules.

ref
reference object (not working)

key
character with key column name.

sparse
logical should only fails be stored in the db?

compute
logical if TRUE the check stores a temporary table in the database.

...
passed through to compute(), if compute is TRUE

See also

Examples

# create a table in a database
income <- data.frame(id = letters[1:2], age=c(12,35), salary = c(1000,NA))
con <- dbplyr::src_memdb()
tbl_income <- dplyr::copy_to(con, income, overwrite=TRUE)
print(tbl_income)

#> # Source:   table<income> [?? x 3]
#> # Database: sqlite 3.33.0 [:memory:]
#>   id      age salary
#>   <chr> <dbl>  <dbl>
#> 1 a        12   1000
#> 2 b        35     NA

# Let's define a rule set and confront the table with it:
rules <- validator( is_adult   = age >= 18
                  , has_income = salary > 0
                  , mean_age   = mean(age,na.rm=TRUE) > 20
                  )

# and confront!
cf <- confront(tbl_income, rules)
print(cf)

#> Object of class 'tbl_validation'
#> Call:
#>     confront.tbl_sql(tbl = dat, x = x, ref = ref, key = key, sparse = sparse)
#> 
#> Confrontations: 3
#> Fails         : [??] (see `values`)
#> Errors        : 0

# Values (i.e. validations on the table) can be retrieved like in `validate` 
# with`type="matrix"` (simplify = TRUE)
values(cf, type = "matrix")

#> [[1]]
#>      is_adult has_income
#> [1,]    FALSE       TRUE
#> [2,]     TRUE         NA
#> 
#> [[2]]
#>      mean_age
#> [1,]     TRUE
#> 

# But often this seems more handy:
values(cf, type = "tbl")

#> # Source:   lazy query [?? x 3]
#> # Database: sqlite 3.33.0 [:memory:]
#>   is_adult has_income mean_age
#>      <int>      <int>    <int>
#> 1        0          1        1
#> 2        1         NA        1

# We can see the sql code by using `show_query`:
show_query(cf)

#> <SQL>
#> SELECT (`age` - 18.0) >= -1e-08 AS `is_adult`, `salary` > 0.0 AS `has_income`, AVG(`age`) OVER () > 20.0 AS `mean_age`
#> FROM `income`

# identical
show_query(values(cf, type = "tbl"))

#> <SQL>
#> SELECT (`age` - 18.0) >= -1e-08 AS `is_adult`, `salary` > 0.0 AS `has_income`, AVG(`age`) OVER () > 20.0 AS `mean_age`
#> FROM `income`

cf <- confront(tbl_income, rules, key = "id")
print(cf)

#> Object of class 'tbl_validation'
#> Call:
#>     confront.tbl_sql(tbl = dat, x = x, ref = ref, key = key, sparse = sparse)
#> 
#> Confrontations: 3
#> key column    : 'id'
#> Fails         : [??] (see `values`)
#> Errors        : 0

values(cf, type="tbl")

#> # Source:   lazy query [?? x 4]
#> # Database: sqlite 3.33.0 [:memory:]
#>   id    is_adult has_income mean_age
#>   <chr>    <int>      <int>    <int>
#> 1 a            0          1        1
#> 2 b            1         NA        1

# sparse results in db
cf <- confront(tbl_income, rules, sparse=TRUE)
values(cf, type="tbl")

#> # Source:   lazy query [?? x 3]
#> # Database: sqlite 3.33.0 [:memory:]
#>     row rule        fail
#>   <int> <chr>      <int>
#> 1     1 is_adult       1
#> 2     2 has_income    NA