Create a sparse confrontation query

Create a sparse confrontation query. Only errors and missing are stored. This can be useful alternative to confront_tbl() which stores all results of a tbl validation in a table with length(rules) columns and nrow(tbl) rows. Note that the result of this function is a (lazy) query object that still needs to be executed in the database, e.g. with dplyr::collect(), dplyr::collapse() or dplyr::compute().

confront_tbl_sparse(
  tbl,
  x,
  key = NULL,
  union_all = TRUE,
  ...,
  check_rules = TRUE
)

Arguments

tbl
tbl object

x
validator object

key
character of key column (not working)

union_all
if FALSE each rule is a separate query.

...
not used

check_rules
if TRUE it is checked which rules 'work' on the db.

Examples

# create a table in a database
income <- data.frame(age=c(12,35), salary = c(1000,NA))
con <- dbplyr::src_memdb()
tbl_income <- dplyr::copy_to(con, income, overwrite=TRUE)
print(tbl_income)

#> # Source:   table<income> [?? x 2]
#> # Database: sqlite 3.33.0 [:memory:]
#>     age salary
#>   <dbl>  <dbl>
#> 1    12   1000
#> 2    35     NA

# Let's define a rule set and confront the table with it:
rules <- validator( is_adult   = age >= 18
                  , has_income = salary > 0
                  )

# and confront!
cf <- confront(tbl_income, rules)
print(cf)

#> Object of class 'tbl_validation'
#> Call:
#>     confront.tbl_sql(tbl = dat, x = x, ref = ref, key = key, sparse = sparse)
#> 
#> Confrontations: 2
#> Fails         : [??] (see `values`)
#> Errors        : 0

# Values (i.e. validations on the table) can be retrieved like in `validate` 
# with`type="list"`
values(cf, type = "list")

#> $is_adult
#> [1] FALSE  TRUE
#> 
#> $has_income
#> [1] TRUE   NA
#> 

# But often this seems more handy:
values(cf, type = "tbl")

#> # Source:   lazy query [?? x 2]
#> # Database: sqlite 3.33.0 [:memory:]
#>   is_adult has_income
#>      <int>      <int>
#> 1        0          1
#> 2        1         NA

# We can see the sql code by using `show_query`:
show_query(cf)

#> <SQL>
#> SELECT (`age` - 18.0) >= -1e-08 AS `is_adult`, `salary` > 0.0 AS `has_income`
#> FROM `income`

# identical
show_query(values(cf, type = "tbl"))

#> <SQL>
#> SELECT (`age` - 18.0) >= -1e-08 AS `is_adult`, `salary` > 0.0 AS `has_income`
#> FROM `income`